+++ src/display/text.c
@@ -31,10 +31,10 @@
 void render_text(fb_buf_t *buf, const char *text, int x, int y, bool v_centre, colour16_t colour)
 {
     if (x < 0)
-        x += (int)buf->size_x - (blit32_ADVANCE * (int)strlen(text));
+        x += buf->size_x - (blit32_ADVANCE * (int)strlen(text));
 
     if (y < 0)
-        y += (int)buf->size_y - blit32_HEIGHT;
+        y += buf->size_y - blit32_HEIGHT;
 
     if (v_centre)
         y -= blit32_HEIGHT / 2;
@@ -43,8 +43,8 @@ void render_text(fb_buf_t *buf, const char *text, int x, int y, bool v_centre, c
         (blit_pixel *)buf->buf,
         colour,
         1,
-        (int)buf->size_x,
-        (int)buf->size_y,
+        buf->size_x,
+        buf->size_y,
         blit_Clip,
         x,
         y,
